BD - Implement classes for nosql capability
