+++ framework/entity/src/main/java/org/apache/ofbiz/entity/GenericNoSqlDelegator.java
@@ -1,0 +1,5 @@
+package org.apache.ofbiz.entity;
+
+public abstract class GenericNoSqlDelegator implements NoSqlDelegator {
+
+}
